Update virtual-machines-deploy-rmtemplates-powershell.md
Remove Section to Upload VHD.  The section instructions do not work.  It refers to a dependent blog post to create the storage account and upload the VHD.  That blog post, does not use resource manager to create the storage account. (See Windows Directions at: https://azure.microsoft.com/en-us/documentation/articles/virtual-machines-create-upload-vhd-windows-server/)  A storage account created with these instructions can no longer be used to create a VM using Resource Manager.  This is an unsupported scenario.  It gives an error message of:  'Please use a storage account managed by Storage Resource Provider. Use of guruarm is not supported. The uploadVHD part of this script should be done in a separate post and fixed so it does not rely on another post.
  
Also, Changed the name of the ResourceGroups and Deployment  in the script (for the second section).  If you follow the blog post in sequence it will not work because when you get to creating the second section (which was UploadVHD but is now Multiple VMs) using the same ResourceGroup Name and Deployment Name it tries to do an update instead of creating the new systems.  Changing the default names will fix this behavior.